refactor: move getEtag utility to fetchWithEtag and update import paths
